Added padder (WIP)
Added a padder so that in case your sequence length is less than the
input size, the model doesn't fail. Rather, it will adapt to the length
of the model. Note, if the input is size (batch, n, c), and the
Linformer has `input_size` set to N, the runtime will be O(Nk), not
O(nk).