Added causal mask, encdec module, benchmarking
Added the option to generate an upper triangular mask, just like in the
original attention is all you need paper. Furthermore, started working
on an encoder/decoder module. This just trains your model on both an
encoder and a decoder, but in the future, one will be able to add their
own sequences to this and decode stuff by themselves.